KAW-8174 Refactor sliding
